Add support for instances without an external IP address.
This change adds two flags:

1. A `beta-no-external-ip` flag for the `create` and `beta create-gpu`
   commands.
2. A `beta-internal-ip` flag to the `connect` command.

These tell the tool to create an instance with no external IP address
and connect to the instance's internal IP addres, respectively.

Connecting to these instances relies on the `gcloud beta compute ssh`
command in order to use the `--internal-ip` flag it supports.

As such, this should not be merged in/released until after that gcloud
command is moved to the GA version of the command.

This fixes #1844